requirements.txt
pandas-flavor
pandas
sentence_transformers
numpy==1.24.3
umap-learn
annchor
matplotlib
scikit-learn
pytest